Move {json,xml}(source_locationt) to {json,xml}_irep.{h,cpp}
This prepares the move of {json,xml}_expr.{h,cpp} to goto-programs to remove the
dependency of util on langapi.
