+++ src/cbmc/bmc_cover.cpp
@@ -16,11 +16,11 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <goto-checker/bmc_util.h>
 
-#include <util/xml.h>
-#include <util/xml_expr.h>
 #include <util/json.h>
-#include <util/json_stream.h>
 #include <util/json_expr.h>
+#include <util/json_stream.h>
+#include <util/xml_expr.h>
+#include <util/xml_irep.h>
 
 #include <solvers/prop/cover_goals.h>
 #include <solvers/prop/literal_expr.h>
