+++ src/goto-programs/show_goto_functions_xml.cpp
@@ -14,9 +14,9 @@ Author: Thomas Kiley
 #include <iostream>
 #include <sstream>
 
-#include <util/xml_expr.h>
 #include <util/cprover_prefix.h>
 #include <util/prefix.h>
+#include <util/xml_irep.h>
 
 #include <langapi/language_util.h>
