+++ src/goto-programs/xml_goto_trace.cpp
@@ -15,8 +15,9 @@ Author: Daniel Kroening
 
 #include <cassert>
 
-#include <util/xml_expr.h>
 #include <util/symbol.h>
+#include <util/xml_expr.h>
+#include <util/xml_irep.h>
 
 #include <langapi/language_util.h>
