+++ src/pointer-analysis/value_set_analysis.cpp
@@ -11,10 +11,9 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "value_set_analysis.h"
 
-#include <util/prefix.h>
 #include <util/cprover_prefix.h>
-#include <util/xml_expr.h>
-#include <util/xml.h>
+#include <util/prefix.h>
+#include <util/xml_irep.h>
 
 #include <langapi/language_util.h>
