+++ src/util/json_irep.h
@@ -13,9 +13,9 @@ Author: Thomas Kiley, thomas.kiley@diffblue.com
 #define CPROVER_UTIL_JSON_IREP_H
 
 #include "irep.h"
+#include "json.h"
 
-class jsont;
-class json_objectt;
+class source_locationt;
 
 class json_irept
 {
@@ -37,5 +37,7 @@ private:
 
   bool include_comments;
 };
+
+json_objectt json(const source_locationt &);
 
 #endif // CPROVER_UTIL_JSON_IREP_H
