+++ src/util/ui_message.cpp
@@ -12,12 +12,10 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <iostream>
 
 #include "cmdline.h"
-#include "json.h"
-#include "json_expr.h"
+#include "json_irep.h"
 #include "json_stream.h"
 #include "make_unique.h"
-#include "xml.h"
-#include "xml_expr.h"
+#include "xml_irep.h"
 
 ui_message_handlert::ui_message_handlert(
   message_handlert *_message_handler,
