+++ src/util/xml_irep.h
@@ -10,8 +10,10 @@ Author: Daniel Kroening, kroening@kroening.com
 #ifndef CPROVER_UTIL_XML_IREP_H
 #define CPROVER_UTIL_XML_IREP_H
 
+#include "xml.h"
+
 class irept;
-class xmlt;
+class source_locationt;
 
 void convert(
   const irept &irep,
@@ -20,5 +22,7 @@ void convert(
 void convert(
   const xmlt &xml,
   irept &irep);
+
+xmlt xml(const source_locationt &);
 
 #endif // CPROVER_UTIL_XML_IREP_H
